backend/requirements.txt
azure-identity==1.12.0
Flask==2.2.2
langchain==0.0.131
openai==0.27.2
azure-search-documents==11.4.0b3
azure-storage-blob==12.14.1
msrestazure
requests
azure-cognitiveservices-speech==1.22.0